pyproject.toml
[project]
name = "ruine-keepers"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.19.0",
    "aiogram-media-group>=0.5.1",
    "asyncpg>=0.30.0",
    "colorama>=0.4.6",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]
